Mutualize common code between linux and windows code
